Pulled out edit card into its own component to reduce cognitive complexity
